Migrate project to ESP-IDF, update build and CI workflows
