+++ .github/workflows/build.yml
@@ -1,4 +1,4 @@
-name: Build ESP32-C6 Project
+name: Build ESP32-C6 ESP-IDF Project
 
 on:
   push:
@@ -13,46 +13,44 @@ jobs:
     
     strategy:
       matrix:
-        environment: [esp32c6, esp32_c3]
+        idf_version: ["5.1", "5.2"]
+        target: ["esp32c6", "esp32c3"]
     
     steps:
     - uses: actions/checkout@v4
       with:
         submodules: 'recursive'
     
-    - name: Set up Python
-      uses: actions/setup-python@v5
+    - name: ESP-IDF Build
+      uses: espressif/esp-idf-ci-action@v1
       with:
-        python-version: '3.11'
-    
-    - name: Cache PlatformIO
-      uses: actions/cache@v4
-      with:
-        path: |
-          ~/.platformio
-          ~/.cache/pip
-        key: ${{ runner.os }}-pio-${{ hashFiles('**/platformio.ini') }}
-        restore-keys: |
-          ${{ runner.os }}-pio-
-    
-    - name: Install PlatformIO
-      run: |
-        python -m pip install --upgrade pip
-        pip install platformio
-    
-    - name: Build ${{ matrix.environment }}
-      run: |
-        pio run -e ${{ matrix.environment }}
-    
-    - name: Check code size
-      run: |
-        pio run -e ${{ matrix.environment }} -t size
+        esp_idf_version: v${{ matrix.idf_version }}
+        target: ${{ matrix.target }}
+        path: '.'
     
     - name: Upload firmware
       uses: actions/upload-artifact@v4
       with:
-        name: firmware-${{ matrix.environment }}
+        name: firmware-${{ matrix.target }}-idf${{ matrix.idf_version }}
         path: |
-          .pio/build/${{ matrix.environment }}/firmware.bin
-          .pio/build/${{ matrix.environment }}/firmware.elf
+          build/dryad.bin
+          build/dryad.elf
+          build/partition_table/partition-table.bin
+          build/bootloader/bootloader.bin
+          build/flash_args
+        retention-days: 30
+    
+    - name: Create flash command file
+      run: |
+        echo "Flash command for ${{ matrix.target }}:" > flash_instructions.txt
+        echo "esptool.py -p PORT -b 460800 --before default_reset --after hard_reset write_flash @flash_args" >> flash_instructions.txt
+        echo "" >> flash_instructions.txt
+        echo "Or use idf.py:" >> flash_instructions.txt
+        echo "idf.py -p PORT flash" >> flash_instructions.txt
+    
+    - name: Upload flash instructions
+      uses: actions/upload-artifact@v4
+      with:
+        name: flash-instructions-${{ matrix.target }}-idf${{ matrix.idf_version }}
+        path: flash_instructions.txt
         retention-days: 30
