+++ CMakeLists.txt
@@ -1,4 +1,10 @@
-cmake_minimum_required(VERSION 3.16.0)
+# The following lines of boilerplate have to be in your project's
+# CMakeLists in this exact order for cmake to work correctly
+cmake_minimum_required(VERSION 3.16)
+
+# Include ESP-IDF components
 include($ENV{IDF_PATH}/tools/cmake/project.cmake)
-project(project-dryad)
+
+# Set the project name
+project(dryad)
