--- include/Apps/DisplayApp.h
@@ -1,4 +1,0 @@
-// include display
-#pragma once
-
-void display_task(void* pvParameters);
