+++ main/display.h
@@ -1,0 +1,6 @@
+#ifndef DISPLAY_H
+#define DISPLAY_H
+
+void display_task(void* pvParameters);
+
+#endif // DISPLAY_H
