--- platformio.ini
@@ -1,31 +1,0 @@
-; PlatformIO Project Configuration File
-;
-;   Build options: build flags, source filter
-;   Upload options: custom upload port, speed and extra flags
-;   Library options: dependencies, extra library storages
-;   Advanced options: extra scripting
-;
-; Please visit documentation for the other options and examples
-; https://docs.platformio.org/page/projectconf.html
-
-[env:esp32_c3]
-platform = espressif32
-board = adafruit_qtpy_esp32c3
-framework = espidf
-monitor_speed = 115200
-build_flags =
-    -Iinclude/Apps
-build_src_filter =
-    +<main.cpp>
-board_build.partitions = custom_partition.csv
-
-[env:esp32c6]
-platform = espressif32
-board = esp32-c6-devkitc-1
-framework = espidf
-monitor_speed = 115200
-build_flags =
-    -Iinclude/Apps
-build_src_filter =
-    +<main.cpp>
-board_build.partitions = custom_partition.csv
