+++ sdkconfig.defaults
@@ -1,0 +1,33 @@
+# ESP32-C6 Configuration
+CONFIG_IDF_TARGET="esp32c6"
+
+# Serial flasher config
+CONFIG_ESPTOOLPY_FLASHMODE_DIO=y
+CONFIG_ESPTOOLPY_FLASHFREQ_80M=y
+CONFIG_ESPTOOLPY_FLASHSIZE_4MB=y
+
+# Partition Table
+CONFIG_PARTITION_TABLE_CUSTOM=y
+CONFIG_PARTITION_TABLE_CUSTOM_FILENAME="partitions.csv"
+
+# Enable FreeRTOS stats for debugging
+CONFIG_FREERTOS_GENERATE_RUN_TIME_STATS=y
+CONFIG_FREERTOS_USE_STATS_FORMATTING_FUNCTIONS=y
+CONFIG_FREERTOS_USE_TRACE_FACILITY=y
+
+# Log output
+CONFIG_LOG_DEFAULT_LEVEL_INFO=y
+CONFIG_LOG_MAXIMUM_LEVEL_VERBOSE=y
+CONFIG_LOG_COLORS=y
+
+# Main task stack size
+CONFIG_ESP_MAIN_TASK_STACK_SIZE=4096
+
+# Enable panic handler
+CONFIG_ESP_SYSTEM_PANIC_PRINT_REBOOT=y
+
+# ADC Configuration
+CONFIG_ADC_ONESHOT_CTRL_FUNC_IN_IRAM=n
+
+# Disable unnecessary features to save space
+CONFIG_ESP_WIFI_SOFTAP_SUPPORT=n
