--- src/CMakeLists.txt
@@ -1,13 +1,0 @@
-# This file was automatically generated for projects
-# without default 'CMakeLists.txt' file.
-
-FILE(GLOB_RECURSE app_sources ${CMAKE_SOURCE_DIR}/src/*.*)
-
-idf_component_register(
-    SRCS 
-        "main.cpp"
-        "Apps/MoistureSensorApp.cpp"
-        "Apps/DisplayApp.cpp"
-    INCLUDE_DIRS 
-        "../include"
-)
